requirements.txt
affine==2.3.1
aiohttp==3.8.3
aiosignal==1.2.0
asciitree==0.3.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
azure-appconfiguration-provider==1.0.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
contourpy==1.0.5
cors==1.0.1
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
executing==1.1.0
fasteners==0.18
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.3
idna==3.4
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
multidict==6.0.2
numcodecs==0.10.2
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
planetary-computer==0.4.7
prompt-toolkit==3.0.31
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyproj==3.4.0
PySocks==1.7.1
pystac==1.6.1
pystac-client==0.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
rasterio==1.3.2
requests==2.28.1
requests-file==1.5.1
rioxarray==0.12.2
shapely==2.0.0
six==1.16.0
snuggs==1.4.7
stack-data==0.5.1
tldextract==3.4.0
tomli==2.0.1
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xarray==2022.9.0
yarl==1.8.1
zarr==2.13.3
zope.event==4.5.0
zope.interface==5.4.0
